+++ config-ui/src/pages/project/detail/webhooks-panel.tsx
@@ -21,8 +21,8 @@ import { Button, Intent } from '@blueprintjs/core';
 
 import API from '@/api';
 import { Alert, NoData } from '@/components';
-import type { WebhookItemType } from '@/plugins/register/webook';
-import { WebhookCreateDialog, WebhookSelectorDialog, WebHookConnection } from '@/plugins/register/webook';
+import type { WebhookItemType } from '@/plugins/register/webhook';
+import { WebhookCreateDialog, WebhookSelectorDialog, WebHookConnection } from '@/plugins/register/webhook';
 import { operator } from '@/utils';
 
 import type { ProjectType } from '../types';
