+++ config-ui/src/plugins/register/webhook/components/delete-dialog.tsx
