+++ config-ui/src/plugins/register/webhook/components/view-dialog.tsx
